requirements.txt
python-telegram-bot>=20.0
mcstatus>=5.2.0
flask>=2.0.0anyio==4.9.0
APScheduler==3.11.0
asyncio-dgram==2.2.0
cachetools==4.2.2
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mcrcon==0.7.0
mcstatus==12.0.2
python-telegram-bot==22.2
pytz==2025.2
requests==2.32.3
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
standard-imghdr==3.13.0
tornado==6.5.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
